Auth (#12)
* auth

* Made everything async (#10)

Fixed Vote not showing in GetById(Round)
Changed Api response when AddPlayer to GameRoom

# Conflicts:
#	ScrumPoker.Infrastructure/AutoMapper/MappingProfile.cs
#	ScrumPoker.Web/Controllers/GameRoomController.cs
#	ScrumPoker.Web/Controllers/RoundController.cs
#	ScrumPoker.Web/Controllers/VoteController.cs

* cleanup

* code fix

tried to apply getting Token playerID into masterId

* some updates

* added getting playerID out of context token

edited validation for claim rights in service layer for gameRoom/Vote/Player/Round services

* cleanup

* Added ForbiddenException

fixed naming for currentUserId
fixed issue with Round not showing cause no Round ID

* removed cs.orig's

* Removed IAuthService (not used)

added logical whitespace's
reworked exceptions
reworked update of player

* fix on Microsoft.AspNet

* removed unnecessary using's

* added Authorize in controllers

* added logic for clearing votes in round

* filename for action ActionNotAllowedException

+++ ScrumPoker.Business/GameRoomService.cs
@@ -1,5 +1,6 @@
 using ScrumPoker.Business.Interfaces.Interfaces;
 using ScrumPoker.Business.Models.Models;
+using ScrumPoker.Common.ForbiddenExceptions;
 using ScrumPoker.DataAccess.Interfaces;
 
 namespace ScrumPoker.Business;
@@ -8,10 +9,12 @@ namespace ScrumPoker.Business;
 public class GameRoomService : IGameRoomService
 {
     private readonly IGameRoomRepository _gameRoomRepository;
+    private readonly IUserManager _userManager;
 
-    public GameRoomService(IGameRoomRepository gameRoomRepository)
+    public GameRoomService(IGameRoomRepository gameRoomRepository, IUserManager userManager)
     {
         _gameRoomRepository = gameRoomRepository;
+        _userManager = userManager;
     }
 
     public async Task<List<GameRoom>> GetAll()
@@ -26,15 +29,25 @@ public class GameRoomService : IGameRoomService
 
     public async Task<GameRoom> Create(GameRoom gameRoomRequest)
     {
-        var gameRoom = await _gameRoomRepository.Create(gameRoomRequest);
-        await _gameRoomRepository.AddPlayerToRoom(gameRoom.Id, gameRoom.MasterId);
-        var gameRoomResponse = await _gameRoomRepository.GetById(gameRoom.Id);
+        var currentUserId = _userManager.GetCurrentUserId();
+        
+        gameRoomRequest.MasterId = currentUserId;
+        var gameRooms = await _gameRoomRepository.Create(gameRoomRequest);
+        await _gameRoomRepository.AddPlayerToRoom(gameRooms.Id, gameRooms.MasterId);
+        
+        var gameRoomResponse = await _gameRoomRepository.GetById(gameRooms.Id);
 
         return gameRoomResponse;
     }
 
     public async Task<GameRoom> Update(GameRoom gameRoomRequest)
     {
+        var gameRoomDto = await GetById(gameRoomRequest.Id);
+        var currentUserId = _userManager.GetCurrentUserId();
+        
+        if (gameRoomDto.MasterId != currentUserId)
+            throw new ActionNotAllowedException($"User has not rights to Update game room (ID {gameRoomRequest.Id})");
+
         return await _gameRoomRepository.Update(gameRoomRequest);
     }
 
@@ -50,11 +63,23 @@ public class GameRoomService : IGameRoomService
 
     public async Task AddPlayer(int gameRoomId, int playerId)
     {
+        var gameRoomDto = await GetById(gameRoomId);
+        var currentUserId = _userManager.GetCurrentUserId();
+        
+        if (gameRoomDto.MasterId != currentUserId)
+            throw new ActionNotAllowedException($"User has not rights to Update game room (ID {gameRoomId})");
+
         await _gameRoomRepository.AddPlayerToRoom(gameRoomId, playerId);
     }
 
     public async Task RemovePlayer(int gameRoomId, int playerId)
     {
+        var gameRoomDto = await GetById(gameRoomId);
+        var currentUserId = _userManager.GetCurrentUserId();
+        
+        if (gameRoomDto.MasterId != currentUserId)
+            throw new ActionNotAllowedException($"User has not rights to Update game room (ID {gameRoomId})");
+
         await _gameRoomRepository.RemovePlayerById(gameRoomId, playerId);
     }
 }
